requirements.txt
PyQt6<6.6
opencv-python-headless
numpy<2.0.0
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
facenet-pytorch==2.5.3
pillow
